pyproject.toml
[project]
name = "RAG Content API"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-huggingface>=0.3.1",
    "markdown>=3.8.2",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.0",
    "sqlalchemy>=2.0.43",
    "unstructured>=0.18.13",
    "watchdog>=6.0.0",
]
